+++ packages/Bio/src/utils/monomer-lib/web-editor-monomer-of-library.ts
@@ -43,7 +43,8 @@ export class LibraryWebEditorMonomer implements IWebEditorMonomer {
       at = monomerLib.getRS(smiles);
     } else if (!monomer.lib) {
       // missing
-      return new MissingWebEditorMonomer(biotype, symbol);
+      throw new Error('Unexpected missing monomer without .lib');
+      // return new MissingWebEditorMonomer(biotype, symbol);
     } else {
       // broken
       return new BrokenWebEditorMonomer(biotype, symbol);
@@ -58,7 +59,7 @@ export class LibraryWebEditorMonomer implements IWebEditorMonomer {
       monomer[REQ.MONOMER_TYPE],
       at);
 
-    const colors = getMonomerColors(biotype, monomer);
+    const colors = getMonomerColors(biotype, monomer, monomerLib);
     if (colors) {
       res.textcolor = colors?.textcolor;
       res.linecolor = colors?.linecolor;
@@ -69,7 +70,7 @@ export class LibraryWebEditorMonomer implements IWebEditorMonomer {
   }
 }
 
-function getMonomerColors(biotype: HelmType, monomer: Monomer): IMonomerColors | null {
+function getMonomerColors(biotype: HelmType, monomer: Monomer, monomerLib?: IMonomerLibBase): IMonomerColors | null {
   const currentMonomerSchema = 'default';
   let monomerSchema: string = currentMonomerSchema;
 
@@ -77,22 +78,28 @@ function getMonomerColors(biotype: HelmType, monomer: Monomer): IMonomerColors |
   if (monomer.meta && monomer.meta.colors) {
     const monomerColors: { [colorSchemaName: string]: any } = monomer.meta.colors;
     if (!(currentMonomerSchema in monomerColors)) monomerSchema = 'default';
-    let res = monomerColors[monomerSchema];
+    res = monomerColors[monomerSchema];
   }
 
   if (!res) {
     const biotypeColors: { [symbol: string]: string } | undefined = naturalMonomerColors[biotype];
-    const nColor = biotypeColors?.[monomer.symbol];
-    if (nColor)
-      res = {textColor: "#000000", lineColor: "#000000", backgroundColor: nColor};
+    const nColor: string = biotypeColors?.[monomer.symbol];
+    if (nColor) {
+      const nTextColor = DG.Color.toHtml(DG.Color.getContrastColor(DG.Color.fromHtml(nColor)));
+      res = {textColor: nTextColor, lineColor: '#202020', backgroundColor: nColor};
+    }
   }
 
-  const na = monomer[OPT.NATURAL_ANALOG];
-  if (!res && na) {
-    const biotypeColors: { [symbol: string]: string } | undefined = naturalMonomerColors[biotype];
-    const naColor = biotypeColors?.[na];
-    res = {textColor: "#000000", lineColor: "#000000", backgroundColor: naColor ?? "#FFFFFF",};
+  const naSymbol: string | undefined = monomer[OPT.NATURAL_ANALOG];
+  if (!res && naSymbol) {
+    const polymerType = monomer[REQ.POLYMER_TYPE];
+    const naMonomer = monomerLib?.getMonomer(polymerType, naSymbol);
+    if (naMonomer)
+      return getMonomerColors(biotype, naMonomer);
   }
+
+  if (!res)
+    res = {textColor: "#202020", lineColor: "#202020", backgroundColor: "#A0A0A0"};
 
   return !res ? null : {
     textcolor: res.text ?? res.textColor,